fix1

+++ database/migrations/2026_04_30_052135_add_parent_category_id_to_categories_table.php
@@ -9,8 +9,10 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('categories', function (Blueprint $table) {
-            $table->unsignedBigInteger('parent_category_id')->nullable()->after('slug');
-            $table->foreign('parent_category_id')->references('id')->on('categories')->onDelete('set null');
+            if (!Schema::hasColumn('categories', 'parent_category_id')) {
+                $table->unsignedBigInteger('parent_category_id')->nullable()->after('slug');
+                $table->foreign('parent_category_id')->references('id')->on('categories')->onDelete('set null');
+            }
         });
     }
